Ai/requirements.txt
# ============================================
# Hubble AI Engine - Python Dependencies
# ============================================

# --- Core Framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.0
pydantic-settings>=2.1.0

# --- Machine Learning ---
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
onnxruntime>=1.17.0

# --- CLIP (multimodal alignment) ---
open-clip-torch>=2.24.0

# --- LangChain + LangGraph ---
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.30

# --- Observability ---
langsmith>=0.1.0
structlog>=24.1.0

# --- Storage ---
motor>=3.3.0
redis>=5.0.0

# --- Media Processing ---
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0

# --- Utilities ---
python-dotenv>=1.0.0
httpx>=0.27.0
